feat: solved topological sort only after solns
I was doing a mock interview with Joanne Ou while doing this problem. We
realised I was just taking too long at the end, but I found out the
solution was actually meant to be a topological sort solution. It should
have been easily identifiable if I've done more topological sort
problems. I originally did this problem with BFS, but that was
inefficient and could have needed a cache? But I'm not sure entirely
because I never completed the solution in BFS.